feat(depla-stack-prisma-astro): set up entity files

+++ packages/depla/src/lib/extract-archive.ts
@@ -6,11 +6,17 @@ export const extractArchive = async (zip, context) => {
   return await zip.forEach(async (relativePath, file) => {
     const fileObj = zip.file(file.name);
     const isFile = fileObj;
+    const filePath = path.resolve(path.join('./', relativePath));
     if (isFile) {
-      fs.writeFileSync(
-        path.resolve(path.join('./', relativePath)),
-        Buffer.from(await fileObj.async('arraybuffer'))
-      );
+      if (
+        !fs.existsSync(filePath) ||
+        relativePath.indexOf('generated') !== -1
+      ) {
+        fs.writeFileSync(
+          filePath,
+          Buffer.from(await fileObj.async('arraybuffer'))
+        );
+      }
     } else {
       const dirPath = path.resolve(path.join('./', relativePath));
       mkdirp.sync(dirPath);
